move remaining construction files
