+++ isoquant_lib/graph_based_model_construction.py
@@ -18,7 +18,7 @@ from .isoform_assignment import (
     IsoformMatch,
 )
 from .polya_finder import PolyAInfo
-from .transcript_to_gene_joiner import TranscriptToGeneJoiner
+from isoquant_lib.model_construction.transcript_to_gene_joiner import TranscriptToGeneJoiner
 from .model_construction.context import ModelStore, ModelConstructionContext
 # Re-exported for isoquant.py's `from ...graph_based_model_construction import StrandnessReportingLevel`.
 from .model_construction.context import StrandnessReportingLevel  # noqa: F401
