+++ isoquant_lib/model_construction/fl_graph_constructor.py
@@ -15,7 +15,7 @@ from typing import Optional, Tuple
 
 from ..common import TranscriptNaming, cmp, get_exons
 from ..gene_info import TranscriptModel, TranscriptModelType
-from ..intron_graph import TerminalVertex
+from isoquant_lib.model_construction.intron_graph import TerminalVertex
 from ..isoform_assignment import is_matching_assignment, ReadAssignment
 from ..polya_finder import PolyAInfo
 from .context import StrandnessReportingLevel
